added .route to routes
