+++ src/routes/channels/messageDeleteBulk.route.ts
