+++ src/routes/channels/messageGetBulk.route.ts
